Fix installation for fresh DeepStream 7.0 containers
Major installation improvements:
- Auto-detect CUDA version and export CUDA_VER for builds
- Add utils.h header with clamp template function for bbox parsing
- Fix nvdsparsebbox_Yolo.cpp type conversions and add iostream include
- Update nvmsg_conv_pose Makefile with correct DeepStream 7.0 include paths
- Temporarily skip custom msgconv library (DS 7.0 API incompatibility)
- Use default DeepStream msgconv for Kafka message broker
- Fix Makefile syntax (remove tab before error directive)
- Update INSTALL.sh to skip nvmsgconv build and use default handler

Compilation fixes:
- Fixed clamp() function type mismatches (float vs int vs uint)
- Added missing <iostream> include for std::cerr
- Added nvmsgconv and deepstream_schema include paths
- Fixed Makefile indentation issues

Note: Custom nvmsgconv library needs DeepStream 7.0 API update (NvDsPayload
structure changed). Currently using default msgconv - pose data still included
in metadata, just uses standard DeepStream JSON format.

Tested: Complete installation successful in fresh DeepStream 7.0 container
- Parsing library: libnvdsinfer_custom_impl_Yolo_pose.so ✓
- Modified test5 app with pose visualization ✓
- Configuration files with correct absolute paths ✓

+++ src/nvdsinfer_custom_impl_Yolo_pose/nvdsparsebbox_Yolo.cpp
@@ -25,6 +25,7 @@
 
 #include "nvdsinfer_custom_impl.h"
 
+#include <iostream>
 #include "utils.h"
 
 extern "C" bool
@@ -41,15 +42,15 @@ convertBBox(const float& bx1, const float& by1, const float& bx2, const float& b
   float x2 = bx2;
   float y2 = by2;
 
-  x1 = clamp(x1, 0, netW);
-  y1 = clamp(y1, 0, netH);
-  x2 = clamp(x2, 0, netW);
-  y2 = clamp(y2, 0, netH);
+  x1 = clamp(x1, 0.0f, static_cast<float>(netW));
+  y1 = clamp(y1, 0.0f, static_cast<float>(netH));
+  x2 = clamp(x2, 0.0f, static_cast<float>(netW));
+  y2 = clamp(y2, 0.0f, static_cast<float>(netH));
 
   b.left = x1;
-  b.width = clamp(x2 - x1, 0, netW);
+  b.width = clamp(x2 - x1, 0.0f, static_cast<float>(netW));
   b.top = y1;
-  b.height = clamp(y2 - y1, 0, netH);
+  b.height = clamp(y2 - y1, 0.0f, static_cast<float>(netH));
 
   return b;
 }
